+++ settings.py
@@ -26,7 +26,7 @@ APP_KITS = {
             'task2_quiz',
             'task2_tr1',
             'task2_results',
-            'static_portfolio',
+            'task3_tr13',
             'static_portfolio_results',
             'static_portfolio_b',
             'static_portfolio_b_results',
@@ -72,7 +72,7 @@ APP_KITS = {
             'task2_quiz',
             'task2_tr1',
             'task2_results',
-            'static_portfolio',
+            'task3_tr13',
             'static_portfolio_results',
             'static_portfolio_b',
             'static_portfolio_b_results',
@@ -107,7 +107,7 @@ SESSION_CONFIGS = [
                       'task2_tr1',
                       'task2_tr2',
                       'task2_results',
-                      'static_portfolio',
+                      'task3_tr13',
                       'static_portfolio_chp',
                       'static_portfolio_results',
                       'static_portfolio_b',
