--- static_portfolio/templates/static_portfolio/Results.html
@@ -1,14 +1,0 @@
-{% extends "global/Page.html" %}
-{% load otree static %}
-
-{% block title %}
-    Task 3: Results
-{% endblock %}
-
-{% block content %}
-    <p>Your points in this task are {{player.payoff}}.</p>
-
-    {% next_button %}
-{% endblock %}
-
-
