+++ task3_tr13/templates/task3_tr13/Results.html
@@ -1,0 +1,14 @@
+{% extends "global/Page.html" %}
+{% load otree static %}
+
+{% block title %}
+    Task 3: Results
+{% endblock %}
+
+{% block content %}
+    <p>Your points in this task are {{player.payoff}}.</p>
+
+    {% next_button %}
+{% endblock %}
+
+
